Module Project

+++ Hreshikesh_XworkzModule/src/main/java/com/xworkz/module/service/RegisterService.java
@@ -20,7 +20,7 @@ public interface RegisterService {
 
     boolean verifyOtp(String otp);
 
-    long countByEmail(String email);
+    int countByEmail(String email);
 
 
 }
